+++ app/src/main/java/com/agileburo/anytype/ui/page/PageFragment.kt
@@ -105,6 +105,13 @@ open class PageFragment :
                     marks = editable.extractMarks()
                 )
             },
+            onParagraphTextChanged = { id, editable ->
+                vm.onParagraphTextChanged(
+                    id = id,
+                    text = editable.toString(),
+                    marks = editable.extractMarks()
+                )
+            },
             onSelectionChanged = vm::onSelectionChanged,
             onCheckboxClicked = vm::onCheckboxClicked,
             onFocusChanged = vm::onBlockFocusChanged,