+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/pattern/Matcher.kt
@@ -1,0 +1,5 @@
+package com.agileburo.anytype.core_ui.features.page.pattern
+
+interface Matcher<out T> {
+    fun match(text: String): List<T>
+}