+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/pattern/Pattern.kt
@@ -1,0 +1,5 @@
+package com.agileburo.anytype.core_ui.features.page.pattern
+
+enum class Pattern {
+    NUMBERED, DIVIDER, CHECKBOX, H1, H2, H3, QUOTE, TOGGLE, BULLET
+}